[formatting] Command-line switch to generate an addresses-only JSON file for planet

+++ scripts/gen_geojson_venues.py
@@ -1,3 +1,4 @@
+import argparse
 import logging
 import os
 import subprocess
@@ -13,7 +14,8 @@ from openvenues.extract.util import *
 
 logger = logging.getLogger('geojson')
 
-def gen_venues(d):
+
+def gen_venues(d, require_latlon=True):
     for filename in os.listdir(d):
         f = open(os.path.join(d, filename))
         for line in f:
@@ -37,14 +39,14 @@ def gen_venues(d):
                 props = None
 
                 if item_type in (SCHEMA_DOT_ORG_TYPE, RDFA_TYPE):
-                    props = schema_dot_org_props(item, item_type)
+                    props = schema_dot_org_props(item, item_type, require_latlon=require_latlon)
                 elif item_type in (OG_TAG_TYPE, OG_BUSINESS_TAG_TYPE):
                     og_type = item.get('og:type', '').rsplit(':', 1)[-1].strip().lower()
                     if og_type not in OG_PLACE_TYPES and not (og_type == 'article' and OG_WHITELIST_DOMAINS.search(domain)):
                         continue
-                    props = og_props(item, item_type)
+                    props = og_props(item, item_type, require_latlon=require_latlon)
                 elif item_type == VCARD_TYPE:
-                    props = vcard_props(item)
+                    props = vcard_props(item, require_latlon=require_latlon)
 
                 if props:
                     venues.append(props)
@@ -52,8 +54,10 @@ def gen_venues(d):
             if venues:
                 yield url, canonical, venues
 
+
 def midpoint(x1, x2):
     return float(x1 + x2) / 2
+
 
 def main(input_dir, output_dir):
     formatter = logging.Formatter('%(asctime)s %(levelname)s [%(name)s]: %(message)s')
@@ -61,6 +65,12 @@ def main(input_dir, output_dir):
     handler.setFormatter(formatter)
     logger.addHandler(handler)
     logger.setLevel(logging.INFO)
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--no-lat-lon', type=bool, action='store_true', default=False)
+    args = parser.parse_args()
+
+    require_latlon = not args.no_lat_lon
 
     city_names = []
     rtree = RTreeIndex()
@@ -74,19 +84,23 @@ def main(input_dir, output_dir):
     i = 0
 
     for k, v in all_cities['regions'].iteritems():
-       for city, data in v['cities'].iteritems():
-           bbox = data['bbox']
-           rtree.insert(i, (float(bbox['left']), float(bbox['bottom']), float(bbox['right']), float(bbox['top'])))        
-           city_names.append(city)
-           i += 1
+        for city, data in v['cities'].iteritems():
+            bbox = data['bbox']
+            rtree.insert(i, (float(bbox['left']), float(bbox['bottom']), float(bbox['right']), float(bbox['top'])))        
+            city_names.append(city)
+            i += 1
 
-    files = {name: open(os.path.join(output_dir, 'cities', '{}.geojson'.format(name)), 'w') for name in city_names}
-    planet = open(os.path.join(output_dir, 'planet.geojson'), 'w')
+    if require_latlon:
+        files = {name: open(os.path.join(output_dir, 'cities', '{}.geojson'.format(name)), 'w') for name in city_names}
+    
+    planet_name = 'planet.geojson' if require_latlon else 'planet_addresses_only.json'
+
+    planet = open(os.path.join(output_dir, planet_name), 'w')
 
     i = 0
     seen = set()
 
-    for url, canonical, venues in gen_venues(input_dir):
+    for url, canonical, venues in gen_venues(input_dir, require_latlon=require_latlon):
         domain = urlparse.urlsplit(url).netloc.strip('www.')
         for props in venues:
             lat = props.get('latitude')
@@ -95,41 +109,50 @@ def main(input_dir, output_dir):
             props['url'] = url
             street = props.get('street_address')
             name = props.get('name')
-            h = hash((name, street, lat, lon, domain))
-            if lat is not None and lon is not None and h not in seen:
+            if require_latlon:
+                h = hash((name, street, lat, lon, domain))
+            else:
+                h = hash((name, street, domain))
+            if require_latlon and lat is not None and lon is not None and h not in seen:
                 cities = list(rtree.intersection((lon, lat, lon, lat)))
                 if cities:
                     for c in cities:
                         f = files[city_names[c]]
                     f.write(json.dumps(venue_to_geojson(props)) + '\n')
             if h not in seen:
-                planet.write(json.dumps(venue_to_geojson(props)) + '\n')
+                if require_latlon:
+                    venue = venue_to_geojson(props)
+                else:
+                    venue = props
+                planet.write(json.dumps(venue) + '\n')
+
             seen.add(h)
             i += 1
             if i % 1000 == 0 and i > 0:
                 logger.info('did {}'.format(i))
 
-    logger.info('Creating manifest files')
+    if require_latlon:
+        logger.info('Creating manifest files')
 
-    manifest_files = []
+        manifest_files = []
 
-    for k, v in all_cities['regions'].iteritems():
-        for city, data in v['cities'].iteritems():
-            f = files[city]
-            if f.tell() == 0:
-                f.close()
-                os.unlink(os.path.join(output_dir, 'cities', '{}.geojson'.format(city)))
-                continue
+        for k, v in all_cities['regions'].iteritems():
+            for city, data in v['cities'].iteritems():
+                f = files[city]
+                if f.tell() == 0:
+                    f.close()
+                    os.unlink(os.path.join(output_dir, 'cities', '{}.geojson'.format(city)))
+                    continue
 
-            bbox = data['bbox']
-            lat = midpoint(float(bbox['top']), float(bbox['bottom']))
-            lon = midpoint(float(bbox['left']), float(bbox['right']))
+                bbox = data['bbox']
+                lat = midpoint(float(bbox['top']), float(bbox['bottom']))
+                lon = midpoint(float(bbox['left']), float(bbox['right']))
 
-            manifest_files.append({'latitude': lat, 'longitude': lon, 'file': '{}.geojson'.format(city), 'name': city.replace('_', ', ').replace('-', ' ').title()})
+                manifest_files.append({'latitude': lat, 'longitude': lon, 'file': '{}.geojson'.format(city), 'name': city.replace('_', ', ').replace('-', ' ').title()})
 
-    manifest = {'files': manifest_files}
+        manifest = {'files': manifest_files}
 
-    json.dump(manifest, open(os.path.join(output_dir, 'manifest.json'), 'w'))
+        json.dump(manifest, open(os.path.join(output_dir, 'manifest.json'), 'w'))
 
     logger.info('Done!')
